bugfix: use a pointer for GenerationConfig.Temperature
Use *float32 for the Temperature field of GenerationConfig.
Provide a SetTemperature method for the user's convenience.

Previously, GenerationConfig.Temperature was a float32 even though
the proto had a *float32, and we intpreted zero as nil. This meant
it was impossible to set a zero temperature: attempting to do
so would transmit a nil Temperature to the server, which the server
would interpret as the default.

No other fields in GenerationConfig have this problem because none
of the others have a valid zero value.