genai: change construction of full model name
Previously, the name passed to Client.GenerativeModel was converted
to a full model name for the service by prepending "models/" unless
that prefix was already present, in which case the name was used
unmodified.

For a more general approach, we now treat any name with a slash
verbatim, and only prepend "models/" if there is no slash.