genai: fix live test to not require flags
Requiring flags precludes us from running `go test ./...` in this repository.
The model name is open now so we can just place it as a constant in the
file - it's no longer the only model used by the tests either, so it's less
confusing. The API key can be provided with an env var like in the examples.

With this change, `go test ./...` now succeeds (the live test fill fail if
the env var key is not set).

Also fix some test expectations to be more resilient.