genai: add File time fields
Add three fields to `File` that are of type time.Time.

This requires the changes to protoveneer in
https://github.com/googleapis/google-cloud-go/pull/10266,
which also includes changes to the support packages.

Running `go generate` also introduced two additional `File` fields,
Error and Metadata, that we will deal with later.

There are also additions to the discovery client which don't concern
us.