genai: move source of truth of example_test and add code generator
We'd like to automatically extract code snippets from our examples for
the official documentation; the documentation tooling extracts code
snippets delineated with special [START... and [END... markers.

However, we also want our examples to appear on the pkgsite without these
special comments.

Therefore, the source of truth for example_test is moved to an internal
directory from which snippets will be extracted; a tool processes this
file to remove these special comments before placing it in genai/example_test.go